Enable XDBG tests again
