+++ dev/local.env
@@ -33,6 +33,7 @@ export NODE_2_HTTP_ADDRESS="http://localhost:5051"
 # Top Level Options Default Node
 export XMTPD_SIGNER_PRIVATE_KEY=$ANVIL_ACC_1_PRIVATE_KEY
 export XMTPD_SIGNER_PUBLIC_KEY=$ANVIL_ACC_1_PUBLIC_KEY
+export XMTPD_SIGNER_ADDRESS=$ANVIL_ACC_1_ADDRESS
 export XMTPD_MLS_VALIDATION_GRPC_ADDRESS="http://localhost:60051"
 
 # Gateway Settings
