+++ pkg/integration/xdbg_test.go
@@ -4,58 +4,73 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/xmtp/xmtpd/pkg/integration/builders"
+
 	"github.com/stretchr/testify/require"
 )
 
 func TestXDBGRealMLSPayloads(t *testing.T) {
-	t.Skip("Needs fixing")
 	network := MakeDockerNetwork(t)
+	xmtpdAlias := "xmtpd"
+	gatewayAlias := "gateway"
+	anvilAlias := "anvil"
 
-	xmtpdContainer, err := NewXmtpdContainerBuilder(t).
+	_, err := builders.NewAnvilContainerBuilder(t).
+		WithNetwork(network).
+		WithNetworkAlias(anvilAlias).
+		Build(t)
+	require.NoError(t, err)
+
+	wsHost := fmt.Sprintf("ws://%s:8545", anvilAlias)
+	rpcHost := fmt.Sprintf("http://%s:8545", anvilAlias)
+
+	builders.RegisterNode(t, network, rpcHost, xmtpdAlias)
+	builders.EnableNode(t, network, rpcHost, 100)
+
+	_, err = builders.NewXmtpdContainerBuilder(t).
 		WithImage("ghcr.io/xmtp/xmtpd:dev").
 		WithNetwork(network).
+		WithNetworkAlias(xmtpdAlias).
+		WithWsUrl(wsHost).
+		WithRPCUrl(rpcHost).
 		Build(t)
 	require.NoError(t, err)
 
-	name, err := xmtpdContainer.Name(t.Context())
-	require.NoError(t, err)
-
-	target := fmt.Sprintf("http:/%s:5050", name)
-
-	gatewayContainer, err := NewXmtpdGatewayContainerBuilder(t).
+	_, err = builders.NewXmtpdGatewayContainerBuilder(t).
 		WithImage("ghcr.io/xmtp/xmtpd-gateway:dev").
 		WithNetwork(network).
+		WithNetworkAlias(gatewayAlias).
+		WithWsUrl(wsHost).
+		WithRPCUrl(rpcHost).
 		Build(t)
 	require.NoError(t, err)
 
-	gatewayName, err := gatewayContainer.Name(t.Context())
-	require.NoError(t, err)
+	target := fmt.Sprintf("http://%s:5050", xmtpdAlias)
+	gatewayTarget := fmt.Sprintf("http:/%s:5050", gatewayAlias)
 
-	gatewayTarget := fmt.Sprintf("http:/%s:5050", gatewayName)
-
-	err = NewXdbgContainerBuilder().
+	err = builders.NewXdbgContainerBuilder().
 		WithNetwork(network).
 		WithTarget(target).
 		WithGatewayTarget(gatewayTarget).
-		WithGeneratorType(GeneratorTypeIdentity).
+		WithGeneratorType(builders.GeneratorTypeIdentity).
 		WithCount(10).
 		Build(t)
 	require.NoError(t, err)
 
-	err = NewXdbgContainerBuilder().
+	err = builders.NewXdbgContainerBuilder().
 		WithNetwork(network).
 		WithTarget(target).
 		WithGatewayTarget(gatewayTarget).
-		WithGeneratorType(GeneratorTypeGroup).
+		WithGeneratorType(builders.GeneratorTypeGroup).
 		WithCount(10).
 		Build(t)
 	require.NoError(t, err)
 
-	err = NewXdbgContainerBuilder().
+	err = builders.NewXdbgContainerBuilder().
 		WithNetwork(network).
 		WithTarget(target).
 		WithGatewayTarget(gatewayTarget).
-		WithGeneratorType(GeneratorTypeMessage).
+		WithGeneratorType(builders.GeneratorTypeMessage).
 		WithCount(10).
 		Build(t)
 	require.NoError(t, err)
